requirements.txt
airsim
matplotlib
numpy
opencv_python
Pillow
PyYAML
scikit_image
skimage
timm
torch
tqdm